fixes for u/int32_t and INTPTR_WIDTH

+++ include/bits/stdint.h
@@ -3,14 +3,14 @@
 typedef unsigned __int128    uint128_t;
 
 typedef unsigned long long   uint64_t;
-typedef unsigned long        uint32_t;
+typedef unsigned int         uint32_t;
 typedef unsigned short       uint16_t; 
 typedef unsigned char        uint8_t;
 
 typedef __int128             int128_t;
 
 typedef long long            int64_t;
-typedef long                 int32_t;
+typedef int                  int32_t;
 typedef short                int16_t;
 typedef char                 int8_t;
